refactor code jwt auth filter and change test controller
