+++ src/main/java/com/homechoice/security/jwt/JwtAuthFilter.java
@@ -25,7 +25,7 @@ public class JwtAuthFilter extends OncePerRequestFilter {
 
     @Override
     protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain)
-            throws ServletException, IOException, ServletException, IOException {
+            throws ServletException, IOException {
 
         final String token = getTokenFromRequest(request);
